# docs/swagger.yaml
basePath: /
definitions:
...
  model.Product:
    properties:
      amount:
        type: integer
      category:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      image_url:
        type: string
      imageID:
        type: string
      order_id:
        type: integer
      price:
        type: number
      purchased_amount:
        type: integer
      related_products:
        items:
          $ref: '#/definitions/model.Product'
        type: array
      reviews:
        items:
          $ref: '#/definitions/model.Review'
        type: array
      tag:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
      views:
        type: integer
    required:
    - amount
    - category
    - price
    - title
    type: object
  model.Review:
    properties:
      category:
        type: string
      created_at:
        type: string
      id:
        type: integer
      product_id:
        type: integer
      text:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
      username:
        type: string
    required:
    - category
    - text
    type: object
  model.User:
    properties:
      password:
        type: string
      role:
        type: string
      username:
        type: string
    required:
    - password
    - role
    - username
    type: object
...
  v1.errorResponse:
    properties:
      message:
        type: string
    type: object
...
  v1.reviewInput:
    properties:
      category:
        type: string
      text:
        type: string
    required:
    - category
    - text
    type: object
  v1.signInInput:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Simple market API
  title: Market API
  version: "1.0"
paths:
...
  /api/v1/product/{productId}/review:
    post:
      consumes:
      - application/json
      operationId: create-review
      parameters:
      - description: ID of product for review
        in: path
        name: productId
        required: true
        type: integer
      - description: Review content
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.reviewInput'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create review
      tags:
      - review
  /api/v1/product/{productId}/review/{reviewId}:
    delete:
      operationId: delete-review
      parameters:
      - description: ID of product
        in: path
        name: productId
        required: true
        type: integer
      - description: ID of review
        in: path
        name: reviewId
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete review
      tags:
      - review
    put:
      consumes:
      - application/json
      operationId: update-review
      parameters:
      - description: ID of product
        in: path
        name: productId
        required: true
        type: integer
      - description: ID of review
        in: path
        name: reviewId
        required: true
        type: integer
      - description: Review content
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.reviewInput'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update review
      tags:
      - review
...
  /api/v1/user/sign-in:
    post:
      consumes:
      - application/json
      operationId: login
      parameters:
      - description: Username and password
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.signInInput'
      produces:
      - application/json
      responses:
        "200":
          description: token
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.errorResponse'
      summary: Login into market
      tags:
      - user
  /api/v1/user/sign-up:
    post:
      consumes:
      - application/json
      operationId: register
      parameters:
      - description: Account info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "200":
          description: token
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.errorResponse'
      summary: Register in the market
      tags:
      - user
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"